fix: renamed typo `constrainted_strings` to `constrained_strings`
