+++ polyfactory/factories/pydantic_factory.py
@@ -27,7 +27,7 @@ from polyfactory.value_generators.constrained_numbers import (
     handle_constrained_float,
     handle_constrained_int,
 )
-from polyfactory.value_generators.constrainted_strings import (
+from polyfactory.value_generators.constrained_strings import (
     handle_constrained_string_or_bytes,
 )
